+++ store/buildPlayer.js
@@ -43,11 +43,6 @@ var projections = {
     rankings: basic,
     hyperopia: basic
 };
-//Fields to project from Cassandra player caches
-//TODO currently we do live significance check.  Persist it to store so we can project fewer fields?
-var cacheProj = ['account_id', 'match_id', 'player_slot', 'version', 'start_time', 'duration', 'game_mode', 'lobby_type', 'radiant_win'];
-var cacheTable = ['hero_id', 'game_mode', 'skill', 'duration', 'kills', 'deaths', 'assists', 'last_hits', 'gold_per_min', 'parse_status'];
-var cacheFilters = ['pgroup', 'hero_id', 'isRadiant', 'lane_role', 'game_mode', 'lobby_type', 'region', 'patch', 'start_time'];
 //Fields to aggregate on
 //optimize by only aggregating certain columns based on tab
 //set query.js_agg based on this
@@ -70,6 +65,11 @@ var aggs = {
     rankings: basicAggs,
     hyperopia: basicAggs
 };
+//Fields to project from Cassandra player caches
+//TODO currently we do live significance check.  Persist it to store so we can project fewer fields?
+var cacheProj = ['account_id', 'match_id', 'player_slot', 'version', 'start_time', 'duration', 'game_mode', 'lobby_type', 'radiant_win'];
+var cacheTable = ['hero_id', 'game_mode', 'skill', 'duration', 'kills', 'deaths', 'assists', 'last_hits', 'gold_per_min', 'parse_status'];
+var cacheFilters = ['pgroup', 'hero_id', 'isRadiant', 'lane_role', 'game_mode', 'lobby_type', 'region', 'patch', 'start_time'];
 
 function buildPlayer(options, cb)
 {
@@ -100,8 +100,6 @@ function buildPlayer(options, cb)
     //choose fields to project based on tab/filter
     //we need to project everything to build a new cache/toplist, otherwise optimize and do a subset
     queryObj.project = config.ENABLE_PLAYER_CACHE || queryObj.keywords.desc ? everything : projections[info];
-    //fields to project from the Cassandra cache
-    queryObj.cacheProject = Object.keys(queryObj.js_agg).concat(cacheProj).concat(cacheTable).concat(filter_exists ? cacheFilters : []);
     //choose fields to aggregate based on tab
     var obj = {};
     aggs[info].forEach(function(k)
@@ -109,6 +107,8 @@ function buildPlayer(options, cb)
         obj[k] = 1;
     });
     queryObj.js_agg = obj;
+    //fields to project from the Cassandra cache
+    queryObj.cacheProject = Object.keys(queryObj.js_agg).concat(cacheProj).concat(cacheTable).concat(filter_exists ? cacheFilters : []);
     //Find player in db
     console.time("[PLAYER] getPlayer " + account_id);
     getPlayer(db, account_id, function(err, player)
