+++ svc/cacher.js
@@ -2,13 +2,13 @@
  * Worker to handle counting and caching tasks performed when a match is inserted or parsed.
  * All operations in this worker should deal with ephemeral data (can be reconstructed from persistent data stores)
  **/
-var constants = require('./constants');
-var config = require('./config');
-var redis = require('./store/redis');
-var queue = require('./store/queue');
-var queries = require('./store/queries');
-var utility = require('./util/utility');
-var benchmarks = require('./compute/benchmarks');
+var constants = require('../constants');
+var config = require('../config');
+var redis = require('../store/redis');
+var queue = require('../store/queue');
+var queries = require('../store/queries');
+var utility = require('../util/utility');
+var benchmarks = require('../compute/benchmarks');
 var cQueue = queue.getQueue('cache');
 var moment = require('moment');
 var async = require('async');
