+++ svc/requests.js
@@ -2,12 +2,12 @@
  * Worker to process parse requests submitted by users
  **/
 var config = require('../config');
-var utility = require('./util/utility');
-var redis = require('./store/redis');
-var db = require('./store/db');
-var cassandra = config.ENABLE_CASSANDRA_MATCH_STORE_WRITE ? require('./store/cassandra') : undefined;
-var queue = require('./store/queue');
-var queries = require('./store/queries');
+var utility = require('../util/utility');
+var redis = require('../store/redis');
+var db = require('../store/db');
+var cassandra = config.ENABLE_CASSANDRA_MATCH_STORE_WRITE ? require('../store/cassandra') : undefined;
+var queue = require('../store/queue');
+var queries = require('../store/queries');
 var getData = utility.getData;
 var pQueue = queue.getQueue('parse');
 var rQueue = queue.getQueue('request');
