+++ svc/retriever.js
@@ -2,7 +2,7 @@
  * Worker interfacing with the Steam GC.
  * Provides HTTP endpoints for other workers.
  **/
-var config = require('./config');
+var config = require('../config');
 var Steam = require('steam');
 var Dota2 = require('dota2');
 var async = require('async');
