+++ svc/worker.js
@@ -1,23 +1,23 @@
 /**
  * Worker running tasks on timed intervals
  **/
-var config = require('./config');
-var constants = require('./constants');
-var redis = require('./store/redis');
-var queue = require('./store/queue');
-var db = require('./store/db');
-var queries = require('./store/queries');
-var buildSets = require('./store/buildSets');
-var utility = require('./util/utility');
-var getMMStats = require("./util/getMMStats");
+var config = require('../config');
+var constants = require('../constants');
+var redis = require('../store/redis');
+var queue = require('../store/queue');
+var db = require('../store/db');
+var queries = require('../store/queries');
+var buildSets = require('../store/buildSets');
+var utility = require('../util/utility');
+var getMMStats = require("../util/getMMStats");
 var async = require('async');
 var moment = require('moment');
 var fs = require('fs');
 var sql = {};
-var sqlq = fs.readdirSync('./sql');
+var sqlq = fs.readdirSync('../sql');
 sqlq.forEach(function(f)
 {
-    sql[f.split('.')[0]] = fs.readFileSync('./sql/' + f, 'utf8');
+    sql[f.split('.')[0]] = fs.readFileSync('../sql/' + f, 'utf8');
 });
 console.log("[WORKER] starting worker");
 invokeInterval(function doBuildSets(cb)
