+++ views/player/player_counts.jade
@@ -3,7 +3,7 @@ extend player
 block playercontent
     .row
       each val, key in counts
-        .col-md-4
+        .col-md-6
           h3=key.split("_").join(" ")
             small  (#{player.aggData[key].n} matches)
           table.table.table-striped.rankable.small
@@ -21,7 +21,8 @@ block playercontent
                   td.rankable=val2
                   td.rankable
                     +progress_bar(player.aggData[key].win_counts[key2], val2)
-      .col-md-4
+    .row
+      .col-md-6
         h3 Kill Streaks 
           small (#{player.aggData.kill_streaks.n} matches)
         table.table.table-striped.rankable.small
@@ -34,7 +35,7 @@ block playercontent
               tr
                 td #{val} (#{key})
                 td.rankable=player.aggData.kill_streaks.counts[key] || 0
-      .col-md-4
+      .col-md-6
         h3 Multi Kills 
           small (#{player.aggData.multi_kills.n} matches)
         table.table.table-striped.rankable.small
